chore: extract animation logic to `ItwFlippableCard` component
